# Conflicts:
#	frontends/php/app/controllers/CControllerMenuPopup.php
#	sass/stylesheets/sass/screen.scss

+++ templates/db/postgresql/postgresql/pgsql.replication.lag.sql
@@ -10,7 +10,7 @@ BEGIN
 			SELECT
 				CASE WHEN pg_last_wal_receive_lsn() = pg_last_wal_replay_lsn()
 					THEN 0
-					ELSE COALESCE(EXTRACT(EPOCH FROM now() - pg_last_xact_replay_timestamp()), 0)
+					ELSE COALESCE(EXTRACT(EPOCH FROM now() - pg_last_xact_replay_timestamp())::integer, 0)
 				END
 			) T;
 
@@ -19,7 +19,7 @@ BEGIN
 			SELECT
 				CASE WHEN pg_last_xlog_receive_location() = pg_last_xlog_replay_location()
 					THEN 0
-					ELSE COALESCE(EXTRACT(EPOCH FROM now() - pg_last_xact_replay_timestamp()), 0)
+					ELSE COALESCE(EXTRACT(EPOCH FROM now() - pg_last_xact_replay_timestamp())::integer, 0)
 				END
 			) T;
 	END IF;
